refactor: Move implementation of mock to sphinx.ext.autodoc.mock
